updated for PR comments
