+++ changelog.d/20230425_115025_derek_implement_get_run_sc_23288.rst
@@ -1,3 +1,2 @@
 
-* `[sc-23288] <https://app.shortcut.com/globus/story/23288>`_
-  Implemented ``FlowsClient.get_run(...)`` (:pr:`NUMBER`)
+* Implemented ``FlowsClient.get_run(...)`` (:pr:`NUMBER`)
