+++ tests/functional/services/flows/test_run_flow.py
@@ -1,0 +1,13 @@
+import typing as t
+
+from globus_sdk import SpecificFlowClient
+from globus_sdk._testing import load_response
+
+
+def test_run_flow(specific_flow_client_class: t.Type[SpecificFlowClient]):
+    metadata = load_response(SpecificFlowClient.run_flow).metadata
+
+    flow_client = specific_flow_client_class(flow_id=metadata["flow_id"])
+
+    resp = flow_client.run_flow(**metadata["request_params"])
+    assert resp.http_status == 200
